Controlling when the builds are run

+++ doc.sh
@@ -1,42 +1,45 @@
 #!/bin/sh
 
-echo "Starting to update documentation...\n"
+if [[ $TRAVIS_BRANCH == "master" ]] ; then
 
-rm -rf docs/                         ç
-./grailsw doc
+    echo "Starting to update documentation...\n"
 
-echo "Replacing version in documentation...\n"
-version=`cat SpringSecurityRestGrailsPlugin.groovy | grep version | sed -e 's/^.*"\(.*\)"$/\1/g'`
-find target/docs/guide -name "*.html" | xargs sed -e "s/&#123;&#123;VERSION&#125;&#125;/${version}/g" -i
+    rm -rf docs/                         ç
+    ./grailsw doc
 
-rm -rf /tmp/docs/
-mv target/docs /tmp
+    echo "Replacing version in documentation...\n"
+    version=`cat SpringSecurityRestGrailsPlugin.groovy | grep version | sed -e 's/^.*"\(.*\)"$/\1/g'`
+    find target/docs/guide -name "*.html" | xargs sed -e "s/&#123;&#123;VERSION&#125;&#125;/${version}/g" -i
 
-#go to home and setup git
-echo "Configuring git...\n"
-cd $HOME
-git config --global user.email "travis@travis-ci.org"
-git config --global user.name "Travis"
+    rm -rf /tmp/docs/
+    mv target/docs /tmp
 
-#using token clone gh-pages branch
-echo "Cloning...\n"
-git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/alvarosanchez/grails-spring-security-rest.git gh-pages > /dev/null
+    #go to home and setup git
+    echo "Configuring git...\n"
+    cd $HOME
+    git config --global user.email "travis@travis-ci.org"
+    git config --global user.name "Travis"
 
-#go into that directory and copy data we're interested in to that directory
-cd gh-pages
+    #using token clone gh-pages branch
+    echo "Cloning...\n"
+    git clone --quiet --branch=gh-pages https://${GH_TOKEN}@github.com/alvarosanchez/grails-spring-security-rest.git gh-pages > /dev/null
 
-echo "Replacing version in index.html...\n"
-cp index.tmpl index.html
-sed -e "s/{{VERSION}}/${version}/g" -i index.html
+    #go into that directory and copy data we're interested in to that directory
+    cd gh-pages
 
-echo "Moving the generated documentation to the right place...\n"
-rm -rf docs/
-mv /tmp/docs .
+    echo "Replacing version in index.html...\n"
+    cp index.tmpl index.html
+    sed -e "s/{{VERSION}}/${version}/g" -i index.html
 
-#add, commit and push files
-echo "Commiting and pushing...\n"
-git add -f .
-git commit -m "Documentation updated"
-git push -fq origin gh-pages > /dev/null
+    echo "Moving the generated documentation to the right place...\n"
+    rm -rf docs/
+    mv /tmp/docs .
 
-echo "Done updating documentation\n"
+    #add, commit and push files
+    echo "Commiting and pushing...\n"
+    git add -f .
+    git commit -m "Documentation updated"
+    git push -fq origin gh-pages > /dev/null
+
+    echo "Done updating documentation\n"
+fi
